# Optimize large tiled backgrounds by prerendering a few repeats into the texture cache.