--- a/src/texture_cache.rs
+++ b/src/texture_cache.rs
@@ -230,6 +230,7 @@ struct TextureCacheArena {
     pages_rgb8: Vec<TexturePage>,
     pages_rgba8: Vec<TexturePage>,
     alternate_pages_a8: Vec<TexturePage>,
+    alternate_pages_rgba8: Vec<TexturePage>,
 }
 
 impl TextureCacheArena {
@@ -239,6 +240,7 @@ impl TextureCacheArena {
             pages_rgb8: Vec::new(),
             pages_rgba8: Vec::new(),
             alternate_pages_a8: Vec::new(),
+            alternate_pages_rgba8: Vec::new(),
         }
     }
 }
@@ -354,6 +356,9 @@ impl TextureCache {
             }
             (ImageFormat::RGBA8, false) => {
                 (&mut self.arena.pages_rgba8, RenderTargetMode::RenderTarget)
+            }
+            (ImageFormat::RGBA8, true) => {
+                (&mut self.arena.alternate_pages_rgba8, RenderTargetMode::RenderTarget)
             }
             (ImageFormat::RGB8, false) => (&mut self.arena.pages_rgb8, RenderTargetMode::None),
             (ImageFormat::Invalid, false) | (_, true) => unreachable!(),
@@ -557,6 +562,20 @@ impl TextureCache {
                                                unblurred_glyph_item.to_image(),
                                                horizontal_blur_item.to_image()))
             }
+            (AllocationKind::TexturePage, TextureInsertOp::Tile(bytes, stretch_size)) => {
+                let scratch_image_id = self.new_item_id();
+                // TODO(pcwalton): Destroy these!
+                self.allocate(scratch_image_id,
+                              0, 0,
+                              stretch_size.width, stretch_size.height,
+                              ImageFormat::RGBA8,
+                              true);
+                let scratch_item = self.get(scratch_image_id);
+                TextureUpdateOp::Update(
+                    result.uv.x, result.uv.y,
+                    width, height,
+                    TextureUpdateDetails::Tile(bytes, stretch_size, scratch_item.to_image()))
+            }
             (AllocationKind::Standalone, TextureInsertOp::Blit(bytes)) => {
                 TextureUpdateOp::Create(self.texture_target_for_standalone_texture(),
                                         width,
@@ -567,6 +586,10 @@ impl TextureCache {
                                         Some(bytes))
             }
             (AllocationKind::Standalone, TextureInsertOp::Blur(_, _, _)) => {
+                println!("ERROR: Can't blur with a standalone texture yet!");
+                return
+            }
+            (AllocationKind::Standalone, TextureInsertOp::Tile(..)) => {
                 println!("ERROR: Can't blur with a standalone texture yet!");
                 return
             }
@@ -625,6 +648,8 @@ fn texture_create_op(texture_size: u32, levels: u32, format: ImageFormat, mode: 
 pub enum TextureInsertOp {
     Blit(Vec<u8>),
     Blur(Vec<u8>, Size2D<u32>, Au),
+    /// Bytes and stretch size, respectively.
+    Tile(Vec<u8>, Size2D<u32>),
 }
 
 trait FitsInside {