Timings:

* Page with nothing but a 1x1 tiled background — 100. ms → 0.423 ms
  (236x improvement)

* https://miiverse.nintendo.net/guide/ — 3.11 ms → 2.12 ms (47%
  improvement)

* http://en.wikipedia.org/wiki/Servomechanism — 0.951 ms → 0.842 ms
  (13% improvement)